push_swap demo

+++ push_swap/libft/libft.h
@@ -15,8 +15,9 @@
 
 # include <stdlib.h>
 # include <unistd.h>
+# include <limits.h>
 
-int				ft_atoi(const char *str);
+long				ft_atoi(const char *str);
 
 void			ft_bzero(void *s, size_t n);
